Move chunks-related functions to a new file (#10172)
* Move chunks-related functions to a new file

Part of https://github.com/pydata/xarray/issues/10089

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

---------
