+++ xarray/__init__.py
@@ -24,7 +24,6 @@ from xarray.computation.computation import (
     cross,
     dot,
     polyval,
-    unify_chunks,
     where,
 )
 from xarray.conventions import SerializationWarning, decode_cf
@@ -52,6 +51,7 @@ from xarray.core.treenode import (
 from xarray.core.variable import IndexVariable, Variable, as_variable
 from xarray.namedarray.core import NamedArray
 from xarray.structure.alignment import align, broadcast
+from xarray.structure.chunks import unify_chunks
 from xarray.structure.combine import combine_by_coords, combine_nested
 from xarray.structure.concat import concat
 from xarray.structure.merge import Context, MergeError, merge
